// packages/web/src/vendor/a2ui/web_core/schema/client-to-server.d.ts
export declare const A2uiClientActionSchema: z.ZodObject<{
    name: z.ZodString;
    surfaceId: z.ZodString;
    sourceComponentId: z.ZodString;
    timestamp: z.ZodString;
    context: z.ZodRecord<z.ZodString, z.ZodAny>;
}, "strict", z.ZodTypeAny, {
    context: Record<string, any>;
    name: string;
    surfaceId: string;
    timestamp: string;
    sourceComponentId: string;
}, {
    context: Record<string, any>;
    name: string;
    surfaceId: string;
    timestamp: string;
    sourceComponentId: string;
}>;
// ...
export type A2uiClientAction = z.infer<typeof A2uiClientActionSchema>;